+++ CHANGELOG.md
@@ -14,6 +14,7 @@ with the `YY.MINOR.MICRO` scheme.
 - SQL supports `trim` function.
 - SQL supports `to_char`, `to_date` functions.
 - SQL supports `ORDER BY` clause.
+- SQL supports `UNION` operator.
 
 <!--
 - New feature `tier` - a group of instances with own replication factor.